move and rework secret Validation
